+++ scripts/scripts.js
@@ -1,6 +1,8 @@
 import {
   loadHeader,
   loadFooter,
+  buildBlock,
+  decorateBlock,
   decorateButtons,
   decorateIcons,
   decorateSections,
@@ -148,6 +150,7 @@ export async function decorateDMImagesWithRendition(
             availableRenditions = json?.repositoryMetadata?.smartcrops || {};
           }
         } catch (e) {
+          // eslint-disable-next-line no-console
           console.error('Failed to fetch metadata:', e);
         }
 
@@ -246,6 +249,40 @@ export async function decorateDMImagesWithRendition(
 }
 
 /**
+ * check if link text is same as the href
+ * @param {Element} link the link element
+ * @returns {boolean} true or false
+ */
+export function linkTextIncludesHref(link) {
+  const href = link.getAttribute('href');
+  const textcontent = link.textContent;
+
+  return textcontent.includes(href);
+}
+
+/**
+   * Builds fragment blocks from links to fragments
+   * @param {Element} main The container element
+   */
+export function buildFragmentBlocks(main) {
+  main.querySelectorAll('a[href]').forEach((a) => {
+    const url = new URL(a.href);
+    if (linkTextIncludesHref(a) && url.pathname.includes('/fragments/')) {
+      const block = buildBlock('fragment', url.pathname);
+      const parent = a.parentElement;
+      a.replaceWith(block);
+      decorateBlock(block);
+      if (parent.tagName === 'P' && parent.querySelector('.block')) {
+        const div = document.createElement('div');
+        div.className = parent.className;
+        while (parent.firstChild) div.appendChild(parent.firstChild);
+        parent.replaceWith(div);
+      }
+    }
+  });
+}
+
+/**
  * Builds all synthetic blocks in a container element.
  * @param {Element} main The container element
  */
@@ -268,6 +305,7 @@ export function decorateMain(main) {
   decorateButtons(main);
   decorateIcons(main);
   buildAutoBlocks(main);
+  buildFragmentBlocks(main);
   decorateSections(main);
   decorateBlocks(main);
 }
